pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-agent"
version = "0.1.0"
description = "The definitive catalog of AI agent patterns. One definition, any framework."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Hovborg", email = "brian@hovborg.nu" },
]
keywords = [
    "ai-agent",
    "multi-agent",
    "agent-framework",
    "agent-catalog",
    "mcp",
    "llm",
    "orchestration",
    "agentic-ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
crewai = ["crewai>=0.80"]
langgraph = ["langgraph>=0.3"]
openai = ["openai>=1.60"]
anthropic = ["anthropic>=0.45"]
google-adk = ["google-adk>=1.0"]
smolagents = ["smolagents>=1.0"]
all = [
    "crewai>=0.80",
    "langgraph>=0.3",
    "openai>=1.60",
    "anthropic>=0.45",
    "google-adk>=1.0",
    "smolagents>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
multiagent = "multiagent.cli:main"

[project.urls]
Homepage = "https://github.com/Hovborg/multi-agent"
Documentation = "https://github.com/Hovborg/multi-agent/tree/main/docs"
Repository = "https://github.com/Hovborg/multi-agent"
Issues = "https://github.com/Hovborg/multi-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/multiagent"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
